Fix redundant highlight element
The line:

<keyword attribute="Constant" context="#stay" String="At Priority Status Indicators" />

did not have a corresponding list. This passed the KDE validation, but
Pandoc now rejects it.

The values that would be in the list are defined explicitly above.
I believe it had to be done that way because the values have an '@'
symbol which wasn't a simple keyword character.